cleaned stuff up
