+++ base/src/background.js
@@ -1,10 +1,6 @@
-/**
- * handles initialization for beeves compatible webextensions
- * stores extension metadata (beeves.json) when the extension is loaded
- * @todo train the NLU backend for the current extension
- */
-
-import * as BeevesUtils from "./BeevesUtils.js";
+import { dispatch } from "./BeevesDispatcher.js";
+import { updateBeevesMetadata } from "./BeevesMetadataUtils.js";
+import { trainNLUBackend } from "./BeevesNLUtils.js";
 import BeevesNativeRouter from "./BeevesNativeRouter.js";
 
 /**
@@ -24,7 +20,7 @@ export const beevesNativeRouterInstance = new BeevesNativeRouter(
     asr: asrMessage => {
       console.log("HANDLING ASR");
       console.log(asrMessage);
-      BeevesUtils.dispatch({ text: asrMessage.transcription });
+      dispatch({ text: asrMessage.transcription });
       return asrMessage;
     }, // handle speech
     sys: x => {
@@ -33,28 +29,12 @@ export const beevesNativeRouterInstance = new BeevesNativeRouter(
   }
 );
 
-
-/**
- * @description metadata storage handler, maintains objects corresponding to
- * beeves.json files and hotword-extension mapping
- */
 browser.runtime.onMessageExternal.addListener(async function(message, sender) {
-  BeevesUtils.printStorage();
-  BeevesUtils.updateBeevesMetadata(message, sender);
-  await BeevesUtils.timeout(1000);
+  updateBeevesMetadata(message, sender);
+  await timeout(1000);
   await trainNLUBackend(sender);
 });
 
-export async function trainNLUBackend(sender) {
-  let snipsfile = (await BeevesUtils.getBeevesMetadata(sender.id))["snips"];
-  let res = await BeevesUtils.putData(
-    `http://localhost:8337/skill/${sender.id}`,
-    snipsfile
-  );
+export async function timeout(ms) {
+  return new Promise(resolve => setTimeout(resolve, ms));
 }
-
-/**
- * @description metadata storage handler, maintains objects corresponding to
- * beeves.json files and hotword-extension mapping
- * @todo REFACTOR
- */
